feat(WorldService): using world use case instead of world service

+++ common/common-domain/src/main/kotlin/ru/woowy/world/WorldTimeRate.kt
@@ -1,7 +1,7 @@
 package ru.woowy.world
 
 enum class WorldTimeRate(
-    val gameHoursPerRealSecond: Int,
+    val gameHoursPerRealSecond: Long,
 ) {
     SLOW(1),
     NORMAL(2),
